+++ packages/stream_video/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unregistered
+
+✅ Added
+* Added option to configure android audio configuration when initializing `StreamVideo` instance by providing `androidAudioConfiguration` to `StreamVideoOptions`.
+
 ## 0.10.2
 
 ✅ Added